rrr: convert rpc/client to module (rrr.client)
Merges client.hpp (2764 lines) + client.cpp (2127 lines) into the
single module unit src/rrr/rpc/client.cpp (~4810 lines). Exports the
Future / FutureGroup / ClientConnection / Client / ClientPool /
bulk-reconnect surface and the std::hash specialization for
rusty::Arc<ClientConnection>. Largest module in the rrr migration.

The fiber_channel.hpp dependency is included textually in the GMF
because rrr.fiber_channel itself is deferred (clang 22 codegen bug
on cross-module template-method instantiation in module purview).
Once fiber_channel becomes a module, swap the textual include for
`import rrr.fiber_channel;`.

The original client.cpp had `using namespace std;` at TU scope;
re-introduced inside the module purview's impl block to avoid
rewriting hundreds of unqualified `list`/`string` references.

The existing GMF forward-decl of `class Client` in
rrr.load_balancer's module unit (`namespace rrr { class Client; }`)
keeps working under clang 22 even though Client is now exported
from rrr.client — clang's attach-once rule lets the GMF
global-module decl coexist with the imported module's full decl
when the consumer doesn't import rrr.client directly.
